requirements.txt
langchain
langchain-google-genai
langchain-community
serpapi
python-dotenv

